cleaning up integration and adding comments
