+++ pkg/integrations/cloudwatch_exporter/cloudwatch_exporter.go
@@ -1,7 +1,6 @@
 package cloudwatch_exporter
 
 import (
-	"context"
 	"fmt"
 	"github.com/grafana/agent/pkg/integrations"
 	integrations_v2 "github.com/grafana/agent/pkg/integrations/v2"
@@ -9,12 +8,6 @@ import (
 	"time"
 
 	"github.com/go-kit/log"
-	yace "github.com/nerdswords/yet-another-cloudwatch-exporter/pkg"
-	yaceConf "github.com/nerdswords/yet-another-cloudwatch-exporter/pkg/config"
-	yaceLog "github.com/nerdswords/yet-another-cloudwatch-exporter/pkg/logger"
-	yaceModel "github.com/nerdswords/yet-another-cloudwatch-exporter/pkg/model"
-	yaceSess "github.com/nerdswords/yet-another-cloudwatch-exporter/pkg/session"
-	"github.com/prometheus/client_golang/prometheus"
 )
 
 func init() {
@@ -22,14 +15,14 @@ func init() {
 	integrations_v2.RegisterLegacy(&Config{}, integrations_v2.TypeMultiplex, metricsutils.NewNamedShim("cloudwatch_exporter"))
 }
 
+// Config is the configuration for the CloudWatch metrics integration
 type Config struct {
 	STSRegion string          `yaml:"stsRegion"`
 	Discovery DiscoveryConfig `yaml:"discovery"`
 	Static    []StaticJob     `yaml:"static"`
 }
 
-// Discovery Jobs
-
+// DiscoveryConfig configures scraping jobs that will auto-discover metrics dimensions for a given service.
 type DiscoveryConfig struct {
 	ExportedTags TagsPerNamespace `yaml:"exportedTags"`
 	Jobs         []*DiscoveryJob  `yaml:"jobs"`
@@ -38,6 +31,7 @@ type DiscoveryConfig struct {
 // TagsPerNamespace represents for each namespace, a list of tags that will be exported as labels in each metric.
 type TagsPerNamespace map[string][]string
 
+// DiscoveryJob configures a discovery job for a given service.
 type DiscoveryJob struct {
 	InlineRegionAndRoles `yaml:",inline"`
 	InlineCustomTags     `yaml:",inline"`
@@ -45,8 +39,7 @@ type DiscoveryJob struct {
 	Metrics              []Metric `yaml:"metrics"`
 }
 
-// Static Jobs
-
+// StaticJob will scrape metrics that match all defined dimensions.
 type StaticJob struct {
 	InlineRegionAndRoles `yaml:",inline"`
 	InlineCustomTags     `yaml:",inline"`
@@ -55,8 +48,6 @@ type StaticJob struct {
 	Dimensions           []Dimension `yaml:"dimensions"`
 	Metrics              []Metric    `yaml:"metrics"`
 }
-
-// Supporting types
 
 // InlineRegionAndRoles exposes for each supported job, the AWS regions and IAM roles in which the agent should perform the
 // scrape.
@@ -90,15 +81,17 @@ type Metric struct {
 	Period     time.Duration `yaml:"period"`
 }
 
+// Name returns the name of the integration this config is for.
 func (c *Config) Name() string {
 	return "cloudwatch_exporter"
 }
 
-// todo: is this used when there's more than one integration instnace in an agent config
+// todo: fix me
 func (c *Config) InstanceKey(agentKey string) (string, error) {
 	return c.Name(), nil
 }
 
+// NewIntegration creates a new integration from the config.
 func (c *Config) NewIntegration(l log.Logger) (integrations.Integration, error) {
 	exporterConfig, err := ToYACEConfig(c)
 	if err != nil {
@@ -106,94 +99,8 @@ func (c *Config) NewIntegration(l log.Logger) (integrations.Integration, error) 
 	}
 	collector := newCloudwatchCollector(l, exporterConfig)
 
-	l.Log("msg", "creating new cloudwatch integration")
-
 	return integrations.NewCollectorIntegration(
 		c.Name(),
 		integrations.WithCollectors(collector),
 	), nil
-}
-
-// cloudwatchCollector
-type cloudwatchCollector struct {
-	logger       yaceLoggerWrapper
-	sessionCache yaceSess.SessionCache
-	scrapeConf   yaceConf.ScrapeConf
-}
-
-func newCloudwatchCollector(logger log.Logger, conf yaceConf.ScrapeConf) *cloudwatchCollector {
-	loggerWrapper := yaceLoggerWrapper{
-		debug: false,
-		log:   logger,
-	}
-	return &cloudwatchCollector{
-		logger:       loggerWrapper,
-		sessionCache: yaceSess.NewSessionCache(conf, true, loggerWrapper),
-		scrapeConf:   conf,
-	}
-}
-
-func (c *cloudwatchCollector) Describe(desc chan<- *prometheus.Desc) {
-	c.logger.Info("running describe in cloudwatch_exporter")
-}
-
-func (c *cloudwatchCollector) Collect(ch chan<- prometheus.Metric) {
-	c.logger.Info("running collect in cloudwatch_exporter")
-
-	reg := prometheus.NewRegistry()
-	cwSemaphore := make(chan struct{}, cloudWatchConcurrency)
-	tagSemaphore := make(chan struct{}, tagConcurrency)
-	observedMetricLabels := map[string]yaceModel.LabelSet{}
-	yace.UpdateMetrics(
-		context.Background(),
-		c.scrapeConf,
-		reg,
-		metricsPerQuery,
-		labelsSnakeCase,
-		cwSemaphore,
-		tagSemaphore,
-		c.sessionCache,
-		observedMetricLabels,
-		c.logger,
-	)
-
-	// Close concurrency channels
-	close(cwSemaphore)
-	close(tagSemaphore)
-
-	reg.Collect(ch)
-}
-
-// yaceLoggerWrapper is wrapper implementation of yaceLog.Logger, based out of a log.Logger.
-type yaceLoggerWrapper struct {
-	log   log.Logger
-	debug bool
-}
-
-func (l yaceLoggerWrapper) Info(message string, keyvals ...interface{}) {
-	l.log.Log(append([]interface{}{"level", "info", "msg", message}, keyvals...)...)
-}
-
-func (l yaceLoggerWrapper) Debug(message string, keyvals ...interface{}) {
-	l.log.Log(append([]interface{}{"level", "debug", "msg", message}, keyvals...)...)
-}
-
-func (l yaceLoggerWrapper) Error(err error, message string, keyvals ...interface{}) {
-	l.log.Log(append([]interface{}{"level", "error", "msg", message, "err", err}, keyvals...)...)
-}
-
-func (l yaceLoggerWrapper) Warn(message string, keyvals ...interface{}) {
-	l.log.Log(append([]interface{}{"level", "warn", "msg", message}, keyvals...)...)
-}
-
-func (l yaceLoggerWrapper) With(keyvals ...interface{}) yaceLog.Logger {
-	withLog := log.With(l.log, keyvals)
-	return yaceLoggerWrapper{
-		log:   withLog,
-		debug: l.debug,
-	}
-}
-
-func (l yaceLoggerWrapper) IsDebugEnabled() bool {
-	return l.debug
 }
